requirements.txt
pyngrok
streamlit
peft
bitsandbytes
pyngrok
langchain
langchain_community
langchain_core
langchain_huggingface
faiss-cpu
sentence-transformers
pypdf
gdown